chore: moved localStorageManager and cookieStorageManager to its own fil
