+++ packages/color-mode/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./color-mode"
 export * from "./initialize-color-mode"
 export type { ColorModeOptions } from "./use-color-mode-state"
-export { cookieStorageManager, localStorageManager } from "./color-mode.utils"
-export type { StorageManager } from "./color-mode.utils"
+export * from "./storage-manager"
